feat: added user feelings
