+++ src/app/controllers/UserController.js
@@ -229,14 +229,6 @@ class UserController {
         return response.sendStatus(200);
     }
 
-    async getFeelings(request, response) {
-        const { email, sentimento } = request.body;
-
-        UserRepository.addFeeling(email, sentimento) // FRONT DEFINIR SE O EMAIL É "ANONIMO" OU O EMAIL REAL!!! ---------
-
-        return response.sendStatus(200);
-    }
-
     async logOut(request, response) {
         /*
             #swagger.tags = ['user']
