+++ src/app/repositories/UserRepository.js
@@ -63,18 +63,6 @@ class UserRepository {
         return updatedEmailVerification;
     }
 
-    async addFeeling(email, sentimento) {
-        const query = {
-            name: "update-feeling-id",
-            text: "INSERT INTO sentimento(email, sentimento) VALUES($1, $2)", // Precisa de coluna pra verified email em usuario, return 0 ou 1
-            values: [email, sentimento],
-        };
-
-        const addFeeling = await db.query(query);
-
-        return addFeeling;
-    }
-
     async findVerifiedEmail(email) {
         const query = {
             name: "fetch-email-id",
@@ -111,42 +99,7 @@ class UserRepository {
         return role[0].cargo;
     }
 
-    async findCoordClass(nome, coordenador_id) {
-        const query = {
-            nome: "fetch-class-id",
-            text: "SELECT id FROM turma WHERE nome = $1 AND coordenador_id = $2",
-            values: [nome, coordenador_id],
-        }
-
-        const id = await db.query(query);
-
-        return id[0].id;
-    }
-
-    async findClassId(classId) {
-        const query = {
-            nome: "find-classroom",
-            text: "SELECT EXISTS (SELECT id FROM turma WHERE id = $1)::bool",
-            values: [classId],
-        }
-
-        const id = await db.query(query);
-
-        return id[0].exists;
-    }
-
-    async checkIfClass() {
-        const query = {
-            nome: "check-classroom",
-            text: "select EXISTS (SELECT * FROM turma)::int",
-        }
-
-        const check = await db.query(query);
-
-        return check;
-    }
-
-    async createUser({ email, nome, senha, cargo }) {
+    async createUser(email, nome, senha, cargo) {
         const hashPass = await bcrypt.hash(senha, 10);
 
         const query = {
@@ -161,54 +114,6 @@ class UserRepository {
         return rows;
     }
 
-    async createClass({ userID, nome, periodo }) {
-
-        const query = {
-            name: "create-class",
-            text: "INSERT INTO turma(coordenador_id, nome, periodo) VALUES($1, $2, $3)",
-            values: [userID, nome, periodo],
-        };
-
-        const rows = await db.query(query);
-
-        return rows;
-    }
-
-    async createStudent({ classId, userId }) {
-        const query = {
-            nome: "create-student",
-            text: "INSERT INTO turma_usuario(turma_id, usuario_id) VALUES($1, $2)",
-            values: [classId, userId],
-        };
-
-        const rows = await db.query(query);
-
-        return rows;
-    }
-
-    async removeStudent({ classId, userId }) {
-        const query = {
-            nome: "remove-student",
-            text: "DELETE FROM turma_usuario WHERE turma_id = $1 AND usuario_id = $2",
-            values: [classId, userId]
-        }
-
-        const rows = await db.query(query);
-
-        return rows;
-    }
-
-    async deleteClass(id) {
-        const query = {
-            name: "delete-class",
-            text: "DELETE FROM turma where turma_id = $1",
-            values: [id],
-        };
-
-        const rows = await db.query(query);
-
-        return rows;
-    }
 }
 
 module.exports = new UserRepository();
