+++ src/routes.js
@@ -16,7 +16,7 @@ router.get('/', (req, res) => {
 router.get('/user/id', UserController.show);
 router.post('/user/add', UserController.add);
 router.post('/user/login', UserController.checkValidLogin);
-router.post('/user/get_feelings', UserController.getFeelings);
+router.post('/user/get_feelings', ClassController.getFeelings);
 router.get('/user/verify', UserController.sendEmail); // Envio do email
 router.post('/user/verify', UserController.verifyEmail); // Verificar email
 router.get('/user/logout', Middleware.authorization, UserController.logOut);
